Wrapper script for mach.  I've been meaning to make this for a while.
It is a lot simpler in scope than it probably looks.

Like most of my build/configure wrapper scripts, it first searches for
the source root by searching the parent directory structure for git/hg
metadata directory.

Then it checks if the current directory has a .mozconfig, and if so,
sets MOZCONFIG so that will get used.  (My plan is to always do this.
I suppose I should actually error out or at least emit a warning if
this problem fails.)

After those two steps, it fires off the `mach` script in the source
root.