Some scripts for tracking memory usage of rustc's build actions.
If I recall correctly, `make-track-memuse-one` is for tracking while
making a single invocation of rustc on a single target.  So your
results may be corrupted if you don't have all of that target's
dependencies in place ahead of time.  I tried to make it warn a little
bit about this scenario where it could.